move the project operator to its own file.
The code is becoming impossible to reason about with everything in
operator.rs
